api: fix endianness handling for non-u8 length fields
Type: fix

When handling length fields larger than u8 in the generated binapi,
the endianness handler was not processing them correctly, leading to
memory corruption. Additionally, if a length field is present,
endianness processing cannot be symmetric, as the length field value
must be read before the host-to-network byte order conversion, and
after the network-to-host conversion. This issue was observed, for
example, in the threads array in sw_interface_tx_placement_details and
sw_interface_set_tx_placement, which use a u32 array_size field to
represent their length. Moreover, there was an error in handling the
context field when sending sw_interface_tx_placement_details to the
client.

Signed-off-by: Ivan Shvedunov <ivan4th@gmail.com>
Change-Id: I2ad5f22a99c5a4eb767121e0abd62613c5a7128c